Separate all QVariant code from PythonSupport.
